OTA A/B update temp
temporary commit, Do not merge

+++ common/arm_serial_mcu.c
@@ -306,8 +306,9 @@ void mcu_get_max9286(void)
 #ifndef QUICKBOOT
 	int i = 0;
 
+    //SUKER warning fixed
 	printf("%s: Send ARM to MCU: cmd:0x%02x, data:0x%02x, size:%lu\n",
-		__func__, cmd, s_cdata[0], sizeof(s_cdata));
+               __func__, cmd, s_cdata[0], (long unsigned int)sizeof(s_cdata));
 #endif
 
 	mcu_serial_tx_data(cmd, s_cdata, sizeof(s_cdata));
